// projects/series-movies/src/app/models/enum/news-category.enum.ts
export enum NewsCategoryType {
  HIDE = "HIDE",
  MEDIA = "MEDIA",
  CRITIC_REVIEW = "CRITIC_REVIEW",
  TELEVISION = "TELEVISION",
  STREAMING = "STREAMING",
  REPORTAGE_MAGAZINE = "REPORTAGE_MAGAZINE",
  NEWS = "NEWS",
  INTERNATIONAL = "INTERNATIONAL",
  VIDEOS_TRAILER = "VIDEOS_TRAILER",
  CINEMA = "CINEMA",
  FUTURE = "FUTURE",
  NEW_THIS_WEEK = "NEW_THIS_WEEK",
  START_OF_SERIES = "START_OF_SERIES",
  START_OF_SEASON = "START_OF_SEASON",
  LISTS = "LISTS",
}